Move files into features/send
