+++ apps/desktop-wallet/src/features/send/AssetAmountsInput.tsx
@@ -39,7 +39,7 @@ import SelectOptionToken from '@/components/Inputs/SelectOptionToken'
 import Truncate from '@/components/Truncate'
 import { useMoveFocusOnPreviousModal } from '@/modals/ModalContainer'
 import ModalPortal from '@/modals/ModalPortal'
-import InputsSection from '@/modals/SendModals/InputsSection'
+import InputsSection from '@/features/send/InputsSection'
 import { Address } from '@/types/addresses'
 import { AssetAmountInputType } from '@/types/assets'
 import { onEnterOrSpace } from '@/utils/misc'
