+++ apps/desktop-wallet/src/features/send/CheckAmountsBox.tsx
