+++ apps/desktop-wallet/src/features/send/CheckFeeLockTimeBox.tsx
@@ -23,7 +23,7 @@ import styled from 'styled-components'
 import Amount from '@/components/Amount'
 import Box from '@/components/Box'
 import HorizontalDivider from '@/components/Dividers/HorizontalDivider'
-import InfoRow from '@/modals/SendModals/InfoRow'
+import InfoRow from '@/features/send/InfoRow'
 import { formatDateForDisplay } from '@/utils/misc'
 
 interface CheckFeeLockTimeBoxProps {
