+++ apps/desktop-wallet/src/features/send/InputsSection.tsx
