+++ apps/desktop-wallet/src/features/send/SendModal.tsx
@@ -33,43 +33,46 @@ import { useWalletConnectContext } from '@/contexts/walletconnect'
 import useAnalytics from '@/features/analytics/useAnalytics'
 import { closeModal } from '@/features/modals/modalActions'
 import { ModalBaseProp } from '@/features/modals/modalTypes'
-import { useAppDispatch, useAppSelector } from '@/hooks/redux'
-import CenteredModal, { ScrollableModalContent } from '@/modals/CenteredModal'
-import ConsolidateUTXOsModal from '@/modals/ConsolidateUTXOsModal'
-import ModalPortal from '@/modals/ModalPortal'
 import {
   buildCallContractTransaction,
   getCallContractWalletConnectResult,
   handleCallContractSend
-} from '@/modals/SendModals/CallContract'
-import CallContractAddressesTxModalContent from '@/modals/SendModals/CallContract/AddressesTxModalContent'
-import CallContractBuildTxModalContent from '@/modals/SendModals/CallContract/BuildTxModalContent'
-import CallContractCheckTxModalContent from '@/modals/SendModals/CallContract/CheckTxModalContent'
+} from '@/features/send/callContract'
+import CallContractAddressesTxModalContent from '@/features/send/callContract/AddressesTxModalContent'
+import CallContractBuildTxModalContent from '@/features/send/callContract/BuildTxModalContent'
+import CallContractCheckTxModalContent from '@/features/send/callContract/CheckTxModalContent'
 import {
   buildDeployContractTransaction,
   getDeployContractWalletConnectResult,
   handleDeployContractSend
-} from '@/modals/SendModals/DeployContract'
-import DeployContractAddressesTxModalContent from '@/modals/SendModals/DeployContract/AddressesTxModalContent'
-import DeployContractBuildTxModalContent from '@/modals/SendModals/DeployContract/BuildTxModalContent'
-import DeployContractCheckTxModalContent from '@/modals/SendModals/DeployContract/CheckTxModalContent'
-import { AddressesTxModalData, TxData } from '@/modals/SendModals/sendTypes'
-import StepsProgress, { Step } from '@/modals/SendModals/StepsProgress'
+} from '@/features/send/deployContract'
+import DeployContractAddressesTxModalContent from '@/features/send/deployContract/AddressesTxModalContent'
+import DeployContractBuildTxModalContent from '@/features/send/deployContract/BuildTxModalContent'
+import DeployContractCheckTxModalContent from '@/features/send/deployContract/CheckTxModalContent'
 import {
-  buildTransferTransaction,
-  getTransferWalletConnectResult,
-  handleTransferSend
-} from '@/modals/SendModals/Transfer'
-import TransferAddressesTxModalContent from '@/modals/SendModals/Transfer/AddressesTxModalContent'
-import TransferBuildTxModalContent from '@/modals/SendModals/Transfer/BuildTxModalContent'
-import TransferCheckTxModalContent from '@/modals/SendModals/Transfer/CheckTxModalContent'
+  AddressesTxModalData,
+  CallContractTxData,
+  DeployContractTxData,
+  TransferTxData,
+  TxContext,
+  TxData,
+  UnsignedTx
+} from '@/features/send/sendTypes'
+import StepsProgress, { Step } from '@/features/send/StepsProgress'
+import { buildTransferTransaction, getTransferWalletConnectResult, handleTransferSend } from '@/features/send/Transfer'
+import TransferAddressesTxModalContent from '@/features/send/Transfer/AddressesTxModalContent'
+import TransferBuildTxModalContent from '@/features/send/Transfer/BuildTxModalContent'
+import TransferCheckTxModalContent from '@/features/send/Transfer/CheckTxModalContent'
+import { useAppDispatch, useAppSelector } from '@/hooks/redux'
+import CenteredModal, { ScrollableModalContent } from '@/modals/CenteredModal'
+import ConsolidateUTXOsModal from '@/modals/ConsolidateUTXOsModal'
+import ModalPortal from '@/modals/ModalPortal'
 import {
   transactionBuildFailed,
   transactionSendFailed,
   transactionsSendSucceeded
 } from '@/storage/transactions/transactionsActions'
 import { Address } from '@/types/addresses'
-import { CallContractTxData, DeployContractTxData, TransferTxData, TxContext, UnsignedTx } from '@/types/transactions'
 
 export type ConfigurableSendModalProps<PT extends { fromAddress: Address }> = {
   txData?: TxData
