+++ apps/desktop-wallet/src/features/send/StepsProgress.tsx
