+++ apps/desktop-wallet/src/features/send/Transfer/AddressesTxModalContent.tsx
@@ -24,8 +24,8 @@ import { InputFieldsColumn } from '@/components/InputFieldsColumn'
 import { useAddressesWithBalance } from '@/features/addressFiltering/addressFilteringHooks'
 import { useAppSelector } from '@/hooks/redux'
 import { ModalContent } from '@/modals/CenteredModal'
-import AddressInputs from '@/modals/SendModals/AddressInputs'
-import { TransferAddressesTxModalOnSubmitData, TransferTxModalData } from '@/modals/SendModals/sendTypes'
+import AddressInputs from '@/features/send/AddressInputs'
+import { TransferAddressesTxModalOnSubmitData, TransferTxModalData } from '@/features/send/sendTypes'
 import { selectAddressByHash, selectIsStateUninitialized } from '@/storage/addresses/addressesSelectors'
 import { isAddressValid, requiredErrorMessage } from '@/utils/form-validation'
