+++ apps/desktop-wallet/src/features/send/Transfer/CheckTxModalContent.tsx
@@ -24,14 +24,14 @@ import styled from 'styled-components'
 
 import FooterButton from '@/components/Buttons/FooterButton'
 import InfoBox from '@/components/InfoBox'
+import CheckAddressesBox from '@/features/send/CheckAddressesBox'
+import CheckAmountsBox from '@/features/send/CheckAmountsBox'
+import CheckFeeLocktimeBox from '@/features/send/CheckFeeLockTimeBox'
+import CheckModalContent from '@/features/send/CheckModalContent'
+import { CheckTxProps, TransferTxData } from '@/features/send/sendTypes'
 import { useAppSelector } from '@/hooks/redux'
 import CenteredModal from '@/modals/CenteredModal'
 import ModalPortal from '@/modals/ModalPortal'
-import CheckAddressesBox from '@/modals/SendModals/CheckAddressesBox'
-import CheckAmountsBox from '@/modals/SendModals/CheckAmountsBox'
-import CheckFeeLocktimeBox from '@/modals/SendModals/CheckFeeLockTimeBox'
-import CheckModalContent from '@/modals/SendModals/CheckModalContent'
-import { CheckTxProps, TransferTxData } from '@/types/transactions'
 import { formatDateForDisplay } from '@/utils/misc'
 
 const TransferCheckTxModalContent = ({ data, fees, onSubmit }: CheckTxProps<TransferTxData>) => {
