+++ apps/desktop-wallet/src/features/send/callContract/AddressesTxModalContent.tsx
@@ -23,8 +23,8 @@ import FooterButton from '@/components/Buttons/FooterButton'
 import { InputFieldsColumn } from '@/components/InputFieldsColumn'
 import { useAppSelector } from '@/hooks/redux'
 import { ModalContent } from '@/modals/CenteredModal'
-import AddressInputs from '@/modals/SendModals/AddressInputs'
-import { CallContractTxModalData } from '@/modals/SendModals/sendTypes'
+import AddressInputs from '@/features/send/AddressInputs'
+import { CallContractTxModalData } from '@/features/send/sendTypes'
 import { selectAddressByHash, selectAllAddressHashes } from '@/storage/addresses/addressesSelectors'
 
 interface CallContractAddressesTxModalContentProps {
