+++ apps/desktop-wallet/src/features/send/deployContract/CheckTxModalContent.tsx
@@ -21,13 +21,13 @@ import { useTranslation } from 'react-i18next'
 import Box from '@/components/Box'
 import FooterButton from '@/components/Buttons/FooterButton'
 import InfoBox from '@/components/InfoBox'
+import CheckAddressesBox from '@/features/send/CheckAddressesBox'
+import CheckAmountsBox from '@/features/send/CheckAmountsBox'
+import CheckFeeLockTimeBox from '@/features/send/CheckFeeLockTimeBox'
+import CheckModalContent from '@/features/send/CheckModalContent'
+import InfoRow from '@/features/send/InfoRow'
+import { CheckTxProps, DeployContractTxData } from '@/features/send/sendTypes'
 import { useAppSelector } from '@/hooks/redux'
-import CheckAddressesBox from '@/modals/SendModals/CheckAddressesBox'
-import CheckAmountsBox from '@/modals/SendModals/CheckAmountsBox'
-import CheckFeeLockTimeBox from '@/modals/SendModals/CheckFeeLockTimeBox'
-import CheckModalContent from '@/modals/SendModals/CheckModalContent'
-import InfoRow from '@/modals/SendModals/InfoRow'
-import { CheckTxProps, DeployContractTxData } from '@/types/transactions'
 
 const DeployContractCheckTxModalContent = ({ data, fees, onSubmit }: CheckTxProps<DeployContractTxData>) => {
   const { t } = useTranslation()
