+++ apps/desktop-wallet/src/features/send/sendTypes.ts
@@ -1,0 +1,103 @@
+/*
+Copyright 2018 - 2024 The Alephium Authors
+This file is part of the alephium project.
+
+The library is free software: you can redistribute it and/or modify
+it under the terms of the GNU Lesser General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+The library is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+GNU Lesser General Public License for more details.
+
+You should have received a copy of the GNU Lesser General Public License
+along with the library. If not, see <http://www.gnu.org/licenses/>.
+*/
+
+import { AssetAmount } from '@alephium/shared'
+import { node } from '@alephium/web3'
+
+import { Address } from '@/types/addresses'
+
+export interface TransferTxData {
+  fromAddress: Address
+  toAddress: string
+  assetAmounts: AssetAmount[]
+  shouldSweep: boolean
+  gasAmount?: number
+  gasPrice?: string
+  lockTime?: Date
+}
+
+export interface CallContractTxData {
+  fromAddress: Address
+  bytecode: string
+
+  assetAmounts?: AssetAmount[]
+  gasAmount?: number
+  gasPrice?: string
+}
+
+export interface DeployContractTxData {
+  fromAddress: Address
+  bytecode: string
+
+  initialAlphAmount?: AssetAmount
+  issueTokenAmount?: string
+  gasAmount?: number
+  gasPrice?: string
+}
+
+export type TxData = TransferTxData | CallContractTxData | DeployContractTxData
+
+export type TransferTxModalData = PartialTxData<TransferTxData, 'fromAddress'>
+export type CallContractTxModalData = PartialTxData<CallContractTxData, 'fromAddress'>
+export type DeployContractTxModalData = PartialTxData<DeployContractTxData, 'fromAddress'>
+
+export type TransferAddressesTxModalOnSubmitData = PartialTxData<TransferTxData, 'fromAddress' | 'toAddress'>
+
+export type AddressesTxModalData =
+  | TransferAddressesTxModalOnSubmitData
+  | DeployContractTxModalData
+  | CallContractTxModalData
+
+export interface TxPreparation {
+  fromAddress: Address
+  bytecode?: string
+  issueTokenAmount?: string
+  alphAmount?: string
+}
+
+export type PartialTxData<T, K extends keyof T> = {
+  [P in keyof Omit<T, K>]?: T[P]
+} & Pick<T, K>
+
+export type CheckTxProps<T> = {
+  data: T
+  fees: bigint
+  onSubmit: () => void
+}
+
+export type UnsignedTx = {
+  fromGroup: number
+  toGroup: number
+  unsignedTx: string
+  gasAmount: number
+  gasPrice: string
+}
+
+export type TxContext = {
+  setIsSweeping: (isSweeping: boolean) => void
+  sweepUnsignedTxs: node.SweepAddressTransaction[]
+  setSweepUnsignedTxs: (txs: node.SweepAddressTransaction[]) => void
+  setFees: (fees: bigint | undefined) => void
+  unsignedTransaction: UnsignedTx | undefined
+  setUnsignedTransaction: (tx: UnsignedTx | undefined) => void
+  unsignedTxId: string
+  setUnsignedTxId: (txId: string) => void
+  setContractAddress: (contractAddress: string) => void
+  isSweeping: boolean
+  consolidationRequired: boolean
+}
